Run aggregation queries and move into main suite
